新增mybatisplus框架，更方便批量，新增swagger框架，实现接口文档自动化更新，用三层架构重构，减少控制层代码，业务逻辑全部转移到service层

+++ backend/src/main/java/com/open/harmony/entity/Order.java
@@ -1,5 +1,6 @@
 package com.open.harmony.entity;
 
+import com.baomidou.mybatisplus.annotation.TableName;
 import com.fasterxml.jackson.annotation.JsonFormat;
 import lombok.*;
 
@@ -27,11 +28,12 @@ import java.util.Date;
 @NoArgsConstructor
 @AllArgsConstructor
 @ToString
+
 public class Order {
 
     private Integer orderId;
 
-    private BigInteger storeId;
+    private Long storeId;
 
     private Double totalPrice;
